% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linegraph.R
\name{linegraph}
\alias{linegraph}
\title{Visual Builder: Line Graph}
\usage{
linegraph(df)
}
\arguments{
\item{df}{The data frame}
}
\description{
This function takes a data frame and creates a shiny web app
for you to create a lien graph with.
There are three customizable elements: x-axis, y-axis, and 2 colors
In this function, colors are predetermined for a gradient
}
\keyword{plot}
\keyword{scatter}
