% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Visual Builder: Scatter plot}
\usage{
scatterplot(df)
}
\arguments{
\item{df}{The data frame}
}
\description{
This function takes a data frame and creates a shiny web app
for you to create a scatter plot with.
There are four customizable elements: x-axis, y-axis, color, and size
}
\keyword{plot}
\keyword{scatter}
